Fix header visibility of remainder of apps
