+++ Tactility/Private/app/crashdiagnostics/QrHelpers.h
