+++ Tactility/Source/app/files/State.cpp
@@ -1,7 +1,8 @@
-#include "State.h"
+#include "app/files/State.h"
+#include "app/files/FileUtils.h"
+
 #include "kernel/Kernel.h"
 #include "Log.h"
-#include "FileUtils.h"
 #include "Partitions.h"
 #include "hal/SdCard.h"
